+++ src/components/Elements/Spinner/Spinner.js
@@ -48,4 +48,4 @@ Spinner.defaultProps = {
   type: 'circle',
 };
 
-export default Spinner;
+export default React.memo(Spinner);