+++ src/index.js
@@ -4,6 +4,7 @@ export { default as ButtonGroup } from './components/Elements/ButtonGroup';
 export { default as Icon } from './components/Elements/Icon';
 export { default as Spinner } from './components/Elements/Spinner';
 export { default as Breadcrumb } from './components/Elements/Breadcrumb';
+export { default as Pill } from './components/Elements/Pill';
 
 // Forms
 export { default as Toggle } from './components/Forms/Toggle';